+++ execve.c
@@ -239,6 +239,23 @@ static int __ld_linux_version(const char *path, int *major, int *minor)
 }
 
 __attribute__((visibility("hidden")))
+int __ld_linux_has_inhibit_cache_option(const char *path)
+{
+	int ret, maj = 0, min = 0;
+
+	ret = __ld_linux_version(path, &maj, &min);
+	if ((ret == -1) && (errno != ENOTSUP))
+		return -1;
+
+	errno = 0;
+	if (ret == -1)
+		return 1;
+
+	/* --inhibit-cache is supported since glibc 2.16 */
+	return (maj > 2) || ((maj == 2) && (min >= 16));
+}
+
+__attribute__((visibility("hidden")))
 int __ld_linux_has_argv0_option(const char *path)
 {
 	int ret, maj = 0, min = 0;
@@ -1162,7 +1179,7 @@ int execve(const char *path, char * const argv[], char * const envp[])
 {
 	char buf[PATH_MAX], hashbangbuf[PATH_MAX], loaderbuf[PATH_MAX];
 	char hashbang[HASHBANG_MAX], loader[HASHBANG_MAX];
-	char *interparg[14+1] = { NULL }; /*  0 ARGV0
+	char *interparg[15+1] = { NULL }; /*  0 ARGV0
 					   *  1 /lib/ld.so
 					   *  2 LD_LINUX_ARGV1
 					   *  3 --preload
@@ -1172,12 +1189,13 @@ int execve(const char *path, char * const argv[], char * const envp[])
 					   *  7 --argv0
 					   *  8 ARGV0
 					   *  9 --inhibit-rpath
-					   * 10 /usr/lib/lib.so:/lib/lib.so
-					   * 11 /bin/sh
-					   * 12 HASHBANG_ARGV1
-					   * 13 -x
-					   * 14 script.sh
-					   * 15 NULL
+					   * 10 --inhibit-cache
+					   * 11 /usr/lib/lib.so:/lib/lib.so
+					   * 12 /bin/sh
+					   * 13 HASHBANG_ARGV1
+					   * 14 -x
+					   * 15 script.sh
+					   * 16 NULL
 					   */
 	int i, j, argc, ret;
 	char * const *arg;
@@ -1286,7 +1304,8 @@ loader:
 	    (__strncmp(loader, "/lib64/ld") == 0)) {
 		char *argv0, *rpath, *runpath, *ld_preload, *ld_library_path,
 		     *inhibit_rpath, *program = real_path;
-		int has_argv0 = 1, has_preload = 1, has_inhibit_rpath = 0;
+		int has_argv0 = 1, has_preload = 1, has_inhibit_rpath = 0,
+		    has_inhibit_cache = 0;
 		int shift = 1, abi = 0;
 		const char *basename;
 		char ldso[NAME_MAX];
@@ -1300,10 +1319,16 @@ loader:
 
 		/*
 		 * the glibc world supports --argv0 since 2.33, --preload since
-		 * 2.30, and --inhibit-rpath since 2.0.94
+		 * 2.30, --inhibit-cache since 2.16, and --inhibit-rpath since
+		 * 2.0.94
 		 */
 		if (__strncmp(ldso, "linux") == 0) {
 			has_inhibit_rpath = 1;
+
+			has_inhibit_cache =
+				   __ld_linux_has_inhibit_cache_option(loader);
+			if (has_inhibit_cache == -1)
+				return -1;
 
 			has_argv0 = __ld_linux_has_argv0_option(loader);
 			if (has_argv0 == -1)
@@ -1371,6 +1396,8 @@ loader:
 			shift += 2;
 		if (has_inhibit_rpath && inhibit_rpath)
 			shift += 2;
+		if (has_inhibit_cache)
+			shift++;
 		for (j = 0; j < i; j++)
 			interparg[j+shift] = interparg[j];
 
@@ -1407,6 +1434,10 @@ loader:
 			interparg[i++] = "--inhibit-rpath";
 			interparg[i++] = inhibit_rpath;
 		}
+
+		/* Add --inhibit-cache */
+		if (has_inhibit_cache)
+			interparg[i++] = "--inhibit-cache";
 
 		/* Add --argv0 and original argv0 */
 		if (has_argv0) {