+++ CHANGELOG.md
@@ -22,6 +22,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add the environment variable `IAMROOT_VERSION`
 - Run the [dynamic loader][ld.so(8)] without options if its path does not start
   by `/lib/ld` or `/lib64/ld`
+- Use the [dynamic loader][ld.so(8)] option `--inhibit-cache` if
+  [exec][exec(3)]'ing to a [glibc] chroot'ed environment
 
 ### Changed